dataset/requirements.txt
gradio
numpy
fiftyone
pillow
psycopg2-binary
pgvector
tritonclient[grpc]
torchvision
